+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/cascades/MatchCandidate.java
@@ -273,6 +273,7 @@ public interface MatchCandidate {
 
         switch (indexType) {
             case IndexTypes.VALUE:
+            case IndexTypes.VERSION:
                 expandIndexMatchCandidate(index,
                     availableRecordTypeNames,
                     availableRecordTypes,